pretty-finished
would like to polish up some rough edges but it is what it is. very proud of this project, did better than i expected. would do it again. put my heart and soul into this again, it was fun coding this as i had a passion and a vision(i was expecting more stuff from my vision but oh well).